==> 1. Filtering the Event Log.ipynb <==
import pandas as pd
import pm4py

## Importing the data

data = pd.read_csv("event_logs/BPI_Challenge_2019_edited.csv", sep=";", parse_dates=["event time:timestamp"])

data.head()

## Looking at contents of columns in the original event log

There are many unnecessary features that contain very little information.

data['case Company'].value_counts()

pd.crosstab(data['case Company'], data['case Document Type'])

data["case Item Category"].value_counts()

The very first time stamp included in the event log is in 1948. This is a mistake.

data["event time:timestamp"].min()

data["event time:timestamp"].max()

len(data[data["event time:timestamp"]<"2018-01-01"].sort_values("event time:timestamp"))

data[data["event time:timestamp"]<"2018-01-01"].nunique()

## Filtering the data, removing and renaming columns

# only sub-process "3-way match, invoice before GR"
data_filtered = data[data["case Item Category"]=="3-way match, invoice before GR"]
data_filtered = data_filtered[data_filtered['case Document Type']=="Standard PO"]
data_filtered = data_filtered[data_filtered['case Item Type']=="Standard"]

# only useful columns
data_filtered = data_filtered[["case concept:name", "case Purchasing Document", "eventID ", "event concept:name", "event time:timestamp", "event org:resource",
               "event Cumulative net worth (EUR)", "case Vendor", "case Spend classification text", "case Spend area text", "case Sub spend area text"]]

# renaming the columns
data_filtered.columns=["case_id", "case_PO_id", "event_id", "event_name", "event_timestamp", "event_resource", "event_value_EUR", 
                       "case_vendor", "case_PR_NPR", "case_product_type", "case_product_specific"]

# removing all cases occurring before 2018
data_filtered = data_filtered[data_filtered["event_timestamp"]>"2018-01-01"]

data_filtered = data_filtered.reset_index(drop=True)

data_filtered

data_filtered.to_csv("event_logs/filtered_event_log.csv", index=False, sep=";", encoding="utf-8")

## Importing the case complete filtered log from ProM to save as .csv

The filtered event log was then filtered further in ProM, in order to select only cases that have been completed (ending on either "clear invoice" or "delete purchase order").

A second dataset was also created, selecting only completed cases that ended on "clear invoice".

log = pm4py.read_xes('event_logs/filtered_event_log_complete.xes')
complete_cases = pm4py.convert_to_dataframe(log)

complete_cases.head()

complete_cases.columns=["event_id", "event_name", "case_PO_id", "case_product_type", "case_product_specific", "event_value_EUR", "event_resource", "remove", "event_timestamp", "case_vendor", "case_PR_NPR", "case_id"]

complete_cases = complete_cases[["case_id", "case_PO_id", "event_id", "event_name", "event_timestamp", "event_resource", "event_value_EUR", 
                       "case_vendor", "case_PR_NPR", "case_product_type", "case_product_specific"]]

complete_cases.nunique()

complete_cases.groupby("case_id").nth(0)["event_name"].unique()

complete_cases.groupby("case_id").nth(-1)["event_name"].unique()

complete_cases.to_csv("event_logs/filtered_event_log_complete.csv", index=False, sep=";", encoding="utf-8")

---
Creating a log containing only cases that finished on "clear invoice" (not delete purchase order item):

log = pm4py.read_xes('event_logs/filtered_event_log_complete_clear_invoice.xes')
complete_cases_clear_invoice = pm4py.convert_to_dataframe(log)

complete_cases_clear_invoice.columns=["event_id", "event_name", "case_PO_id", "case_product_type", "case_product_specific", "event_value_EUR", "event_resource", "remove", "event_timestamp", "case_vendor", "case_PR_NPR", "case_id"]

complete_cases_clear_invoice = complete_cases_clear_invoice[["case_id", "case_PO_id", "event_id", "event_name", "event_timestamp", "event_resource", "event_value_EUR", 
                       "case_vendor", "case_PR_NPR", "case_product_type", "case_product_specific"]]

complete_cases_clear_invoice.nunique()

complete_cases_clear_invoice.groupby("case_id").nth(-1)["event_name"].unique()

complete_cases_clear_invoice.to_csv("filtered_event_log_complete_clear_invoice.csv", index=False, sep=";", encoding="utf-8")
