==> 2. Pre-Processing the Event Log for Modelling.ipynb <==
## Importing the data

import pandas as pd
import pm4py
import numpy as np
import statsmodels.api as sm
import statsmodels.formula.api as smf
import seaborn as sns
import matplotlib.pyplot as plt

data = pd.read_csv("event_logs/filtered_event_log_complete_clear_invoice.csv", sep=";", parse_dates=["event_timestamp"])

The aim of this project is to use case attributes of the purchase orders to create a model that attempts to predict the throughput time of each purchase order. To achieve this, an appropriate dataset must be created from the filtered event log.

## Creating a table of case attributes for each line item

The PR_NPR, product types, vendor and value are generally consistent across all of the events related to a particular line item.  
The associated resources and occurrence of different events in each line items trace will need to be added after.

data.groupby("case_id").nunique().mean()

Starting with the beginning event time stamp and case attributes:

start_case = data.groupby("case_id").nth(0)[["event_timestamp", "event_value_EUR", "case_vendor", "case_PR_NPR", "case_product_type", "case_product_specific"]]
start_case.rename(columns={"event_timestamp":"start_timestamp"})
start_case.head()

Adding the end event timestamp:

end_case = data.groupby("case_id").nth(-1)["event_timestamp"]
end_case = end_case.rename("end_timestamp")

Calculating the case total throughput for each line item:

li_data = pd.merge(start_case, end_case, left_index=True, right_index=True, how="outer")
li_data = li_data.rename(columns={"event_timestamp":"start_timestamp"})

li_data["throughput"] = li_data["end_timestamp"] - li_data["start_timestamp"]

li_data.head()

Some items have no product types included:

li_data[li_data["case_PR_NPR"].isna()].head(5)

li_data = li_data.dropna(how="any")

### Feature engineering other relevant case attributes

Finding rework rate per case:

rework_activities = ['Change Price', 'Change Quantity', 'Change Delivery Indicator',
                     'Vendor creates debit memo', 'Cancel Invoice Receipt','Change Storage Location', 
                     'Change Currency', 'Reactivate Purchase Order Item', 'Change Approval for Purchase Order', 
                     'Cancel Goods Receipt','Cancel Subsequent Invoice', 'Update Order Confirmation', 
                     'Record Subsequent Invoice', 'Change payment term']

li_data = pd.merge(li_data, data[data["event_name"].isin(rework_activities)].groupby("case_id").count()["event_name"].rename("rework_activities"), 
         how="left", left_index=True, right_index=True)
li_data["rework_activities"] = li_data["rework_activities"].fillna(0)

Was a payment block used? ("Remove payment block" occurred):

cases_w_pb = pd.DataFrame(data[data["event_name"]=="Remove Payment Block"]["case_id"].unique())
cases_w_pb["payment_block"] = True
cases_w_pb = cases_w_pb.set_index(0)

li_data = pd.merge(li_data, cases_w_pb, how="left", left_index=True, right_index=True)
li_data["payment_block"] = li_data["payment_block"].fillna(False)

Automation rate per case:

data["automation"] = data["event_resource"].str[0]
data["automation"] = data["automation"].replace({"u": False, 
                                                 "b": True, 
                                                 "N" : np.nan})

li_data = pd.merge(li_data, data.dropna().groupby("case_id")["automation"].mean()*100, how="left", left_index=True, right_index=True)

Average daily workload (number of cases being worked on) for the duration of the case, per vendor:

data["event_date"] = data["event_timestamp"].dt.date

all_cases = list(li_data.index)
all_workload = list()
for case in all_cases:
    vendor = data[data["case_id"] == case]["case_vendor"].iloc[0]
    only_vendor_data = data[data["case_vendor"] == vendor]
    between_dates = only_vendor_data[only_vendor_data["event_timestamp"].between(li_data.loc[case, :]["start_timestamp"], li_data.loc[case, :]["end_timestamp"])]
    avg_daily_workload = between_dates.groupby("event_date").nunique()["case_id"].mean()
    all_workload.append(avg_daily_workload)

pd.concat([pd.Series(li_data.index), pd.Series(all_workload).rename("avg_vendor_workload")], axis=1).to_csv(case_avg_vendor_workload, index=False)

case_avg_vendor_workload = pd.read_csv("throughput_dataset/case_avg_vendor_workload.csv").set_index("case_id")

li_data = pd.merge(li_data, case_avg_vendor_workload, how="left", left_index=True, right_index=True)

Finished line item level case dataset:

li_data = li_data.reset_index()

li_data.head()

## Creating a table of case attributes for each purchase order

Many of the attributes in the line item dataset are taken from the purchase order level, and thus duplicated across multiple observations.  
It therefore may be viable to group cases by the purchase order instead of line item.

Are all of the line items in one purchase always cleared ("clear invoice") on the same day? The majority yes (6961) but 6.5% are not (486).

pd.DataFrame(data.groupby("case_PO_id").apply(lambda group: group[group["event_name"]=="Clear Invoice"])["event_timestamp"]).reset_index().groupby("case_PO_id").nunique()["event_timestamp"].value_counts()

Within one purchase order document, there is on average 3.5 different line items.   

Attributes that are the same for every line item in each purchase order:
* vendor
* PR_NPR
* product type
* specific product type  
* The process (trace) followed seems to be the same for every line item in one purchase order, with the same resources working on each of the items.

Attributes that differ for every line item in each purchase order, and will need to be averaged:
* line item values

Attributes that will need to be calculated from one of the line items in the purchase order:
* automation rate
* rework rate
* payment block occurred

Attributes that only exist at the purchase order level:
* number of line items in one purchase order 

data.groupby("case_PO_id").nunique().mean()

Creating a case attributes dataframe, from the purchase order level:

Because the cases within a purchase order tend to follow the same process, be worked on by the same resources, etc, one line item case is used as the "representative" of the line items in a purchase order.

li_data["case_PO_id"] = li_data["case_id"].str.split("_", expand=True)[0]

PO_data = li_data.groupby("case_PO_id").nth(0).reset_index()

PO_data.head()

### Feature engineering purchase order level case attributes

PO_data = PO_data.set_index("case_PO_id")

Number of line items in each purchase order:

num_items = data.groupby("case_PO_id").nunique()["case_id"].rename("num_items")
num_items.index = num_items.index.astype("str")

PO_data = pd.merge(PO_data, num_items, how="left", left_index=True, right_index=True)

Average line item value in each purchase order:

avg_value = data.groupby("case_PO_id").apply(lambda group: group[group["event_name"]=="Create Purchase Order Item"])["event_value_EUR"].reset_index().groupby("case_PO_id").mean()["event_value_EUR"]
avg_value.index = avg_value.index.astype("str")

PO_data["event_value_EUR"] = avg_value
PO_data = PO_data.rename(columns={"event_value_EUR":"avg_value_EUR"})

Final purchase order level dataset:

PO_data = PO_data.reset_index()

PO_data.head()

li_data.to_csv("throughput_dataset/line_item_dataset.csv", index=False)

PO_data.to_csv("throughput_dataset/PO_dataset.csv", index=False)

## Pre-processing for modelling

For the later modelling, having too many one-hot encoded features created from the factor levels could cause dimensionality problems.   
Factor level aggregation is explored in the notebook "Exploring the Processed Dataset", and implemented here.

The datasets also need to be formatted so that they can be processed by an algorithm: e.g. dropping non-numerical id columns, etc.

li_X = li_data[["event_value_EUR", "case_vendor", "case_PR_NPR", "case_product_type", "case_product_specific", "rework_activities", "payment_block", "automation", "avg_vendor_workload"]]
li_y = li_data["throughput"]

PO_X = PO_data[["avg_value_EUR", "case_vendor", "case_PR_NPR", "case_product_type", "case_product_specific", "rework_activities", "payment_block", "automation", "avg_vendor_workload", "num_items"]]          
PO_y = PO_data["throughput"]

li_y = li_y.dt.days
PO_y = PO_y.dt.days

"other" levels always start with _ so that it becomes the reference group in the models.

# PR/NPR
li_X = li_X.replace({"case_PR_NPR" : {"NPR":"_NPR", "OTHER":"_NPR"}})

# group product types
product_types = {key: "_other_product_type" for key in li_data["case_product_type"].unique()}
product_types["Packaging"] = "Packaging"
product_types["Sales"] = "Sales"
product_types["Additives"] = "Additives"
li_X = li_X.replace({"case_product_type" : product_types})

# specific product types
specific_types = {key: "_other_specific_type" for key in li_data["case_product_specific"].unique()}
specific_types["Products for Resale"] = "Products for Resale"
specific_types["Labels"] = "Labels"
specific_types["Metal Containers & Lids < 30L"] = "Containers"
specific_types["Plastic Containers & Lids < 30L"] = "Containers"
specific_types["Extenders"] = "Extenders"
li_X = li_X.replace({"case_product_specific" : specific_types})

# vendors
vendors = {key: "_other_vendor" for key in li_data["case_vendor"].unique()}
vendors["vendorID_0120"] = "vendorID_0120"
vendors["vendorID_0104"] = "vendorID_0104"
vendors["vendorID_0171"] = "vendorID_0171"
vendors["vendorID_0136"] = "vendorID_0136"
vendors["vendorID_0106"] = "vendorID_0106"
li_X = li_X.replace({"case_vendor" : vendors})

li_X = li_X.replace({"payment_block" : {True:1, False:0}})

li_X

# PR/NPR
PO_X = PO_X.replace({"case_PR_NPR" : {"NPR":"_NPR", "OTHER":"_NPR"}})

# group product types
product_types = {key: "_other_product_type" for key in PO_data["case_product_type"].unique()}
product_types["Packaging"] = "Packaging"
product_types["Sales"] = "Sales"
product_types["Additives"] = "Additives"
PO_X = PO_X.replace({"case_product_type" : product_types})

# specific product types
specific_types = {key: "_other_specific_type" for key in PO_data["case_product_specific"].unique()}
specific_types["Products for Resale"] = "Products for Resale"
specific_types["Labels"] = "Labels"
specific_types["Metal Containers & Lids < 30L"] = "Containers"
specific_types["Plastic Containers & Lids < 30L"] = "Containers"
specific_types["Extenders"] = "Extenders"
PO_X = PO_X.replace({"case_product_specific" : specific_types})

# vendors
vendors = {key: "_other_vendor" for key in PO_data["case_vendor"].unique()}
vendors["vendorID_0120"] = "vendorID_0120"
vendors["vendorID_0104"] = "vendorID_0104"
vendors["vendorID_0171"] = "vendorID_0171"
vendors["vendorID_0136"] = "vendorID_0136"
vendors["vendorID_0106"] = "vendorID_0106"
PO_X = PO_X.replace({"case_vendor" : vendors})

PO_X = PO_X.replace({"payment_block" : {True:1, False:0}})

Comparing the full versus new modelling dataset:

PO_model_data = pd.concat([PO_X, PO_y], axis=1)
li_model_data = pd.concat([li_X, li_y], axis=1)

PO_data

PO_model_data

PO_model_data.to_csv("throughput_dataset/PO_modelling_dataset.csv", index=False)
li_model_data.to_csv("throughput_dataset/line_item_modelling_dataset.csv", index=False)
