==> 4. Modelling Throughput.ipynb <==
## Importing the data

import pandas as pd
import pm4py
import numpy as np
import statsmodels.api as sm
import statsmodels.formula.api as smf
import seaborn as sns
import matplotlib.pyplot as plt

PO_model_data = pd.read_csv("throughput_dataset/PO_modelling_dataset.csv")
li_model_data = pd.read_csv("throughput_dataset/line_item_modelling_dataset.csv")

package_v = PO_model_data[PO_model_data["case_vendor"]=="vendorID_0106"]
variety_v = PO_model_data[PO_model_data["case_vendor"]=="vendorID_0171"]

## Linear regression model on the full data

All variables, line item level:

model_li = smf.ols(formula="throughput ~ event_value_EUR + rework_activities + payment_block + automation + avg_vendor_workload\
                + C(case_PR_NPR) + C(case_product_type) + C(case_product_specific) + C(case_vendor)",
               data=li_model_data).fit() 
model_li.summary()

---
All variables, purchase order level:

model_PO = smf.ols(formula='throughput ~ avg_value_EUR + rework_activities + payment_block + automation + avg_vendor_workload + num_items\
                  + C(case_product_type, Treatment("_other_product_type")) + C(case_product_specific, Treatment("_other_specific_type")) + C(case_vendor)',
               data=PO_model_data).fit() 
model_PO.summary()

----
The fit of the model is extremely poor in both cases.

Looking at the p-values for all of the features in the purchase order model:

pd.DataFrame(np.round(model_PO.pvalues, 3)).reset_index().rename(columns={0:"p-value", "index":"feature name"})\
.merge(pd.DataFrame(np.round(model_PO.params, 3)).reset_index().rename(columns={0:"coefficients", "index":"feature name"}), left_on="feature name", right_on="feature name")

## Linear regression model for two specific vendors

Modelling vendor 0171:

model = smf.ols(formula="throughput ~ avg_value_EUR + rework_activities + payment_block + automation + avg_vendor_workload + num_items\
                + C(case_product_type) + C(case_product_specific)",
               data=variety_v).fit() 
model.summary()

Modelling vendor 0106:

model = smf.ols(formula="throughput ~ avg_value_EUR + rework_activities + payment_block + automation + avg_vendor_workload + num_items\
                + C(case_product_type) + C(case_product_specific)",
               data=package_v).fit() 
model.summary()
